RPA Paycodes (#2006)
This implements RPA Re-usable payment address (AKA Paycodes) with a CLI.

Note this only implements the basic functions of generating a paycode, a transaction,
and extracting private key. Other functions such as storing the key in the keystore will
be implemented in the next stages.

This is a continuation of https://github.com/Electron-Cash/Electron-Cash/pull/2001.  
It was redone on a new branch to avoid cherry picked commits from master.

It incorporates feedback from @cculianu including:

* Avoiding changing cashaddr module and instead use a new rpaaddr module.
* Avoiding changing default behavior in libsecp256k1 by passing null value rather than empty byte string for ndata
* Pythonic looping syntax
* spaces after commas in code
* Avoiding literal "0" for cashaddr and instead use Address.FMT_CASHADDR