Fix to keyserver.py sign_transaction for Flipstarter backwards compat
The recent RPA Paycodes commit (#2006) ended up modifying the
Transaction class API slightly by adding a new kwarg, 'ndata'. This
confused the Flipstarter plugin which inherits from this class,
reimplementing `sign()`.  This broke basically all extant Flipstarter
plugin installs.

Rather than force users to upgrade to a new flipstarter, which is a bit
like herding cats, we decided to work around the situation ourselves in
Electron Cash.

The fix is to detect, at runtime, whether the class in question
reimplements `sign()` and lacks the `ndata` kwarg, and if so, not pass
that arg.

Tested by and signed-off by imaginaryusername