[fusion] fix race during stopping
If we use the .status to check whether a fusion is done, then
there is a race and potentially that fusion still has coins frozen.
During setting up autofusions this could mean we start a new fusion
without having a full view of coins, and during wallet shutdown
it could mean we close a wallet with some coins still marked frozen.

Using .is_alive() is okay because we are guaranteed that all fusions
have been .start()ed immediately after creation / launching.