always use MOST_RELAXED if oneserver is set
if oneserver is set then the user is making a deliberate choice to use a
(potentially less performant and/or privately run) server. as we see in
issue #6625 these are often cases where the user has a large wallet and
therefore may need a longer timeout for processing requests. The user
should not have to care about whether auto_connect is set in order to
get a longer timeout