NetworkDialog: Fix glitches when restoring selection on update()
It was observed that now the network dialog is glitchy -- if you select
a server it either of the two panes (Overview, Server list), it behaves
glitchy and it either doesn't restore your selection or it spazzes and
scrolls the server list for you when it does.

This commit fixes that.

This commit also closes issue #2069.