Fiat display: Unconditionally disabled for the testnets
It may be a phishing/scam vector to offer fiat display for the testnets,
since this could lead novice users to believe that testnet coins have
some monetary value.

In this commit we unconditionally disable fiat display and the
exchange_rate updater if on one of the testnets: TestNet, TestNet4, or
ScaleNet (also if on TaxCoinNet since we don't have accurate exchange
rate info for that).

The Qt GUI has been updated such that the Fiat tab in preferences says
"Fiat display is not supported on this chain." if you are on one of
those networks.

Also all fiat information will never appear in the Qt UI if on a testnet
(or on taxcoin), regardless of what the config object has as a setting.

In this way it is hoped that confusion for novice users is minimized.
(And also the fiat display on a testnet is pretty much not very useful).