AppImage: Build all binaries except PyQt5
This changes the AppImage build to build all binaries instead of
fetching wheels from PyPi, except for `PyQt5`. `PyQt5` is complex to
build and takes a while, so we exclude it. For compiling `cryptography`
we need to add a rust compiler to the AppImage build process.

This needs `pysatochip>=0.11.4` because earlier versions do not compile
correctly due to missing README.md.