NSIS: Ensure the process is not running when (un)installing
Installing when the process is running can have unintended effects:

See https://github.com/spesmilo/electrum/issues/6748

To address this we disallow (un)installing as long as the process is
running. To detect that the process is we try to open the .exe file
with write access (append, so we don't truncate the file) and if that
fails we know the process is running.