Remove CPFP (child pays for parent)
This feature is going away from BCH very soon with the May 15th 2021
network upgrade.  As such, we remove it now.  It hardly was used anyway
on BCH so nobody will miss it if it's simply gone right away without a
transition period.

Miners running BCHN 23.0.0 don't have CPFP code anymore anyway when
generating blocks so.. it's a good idea to remove this ASAP as well.