Add "spend only fused coins" to Send tab
We add a new checkbox to the send tab, "Spend only fused coins", which
only appears if CashFusion is loaded and if the wallet in question can
fuse.

If checked, we do something similar to the CashShuffle plugin -- we
refuse to spend coins that aren't yet fused. We also try to "force-add"
coins that live on the same address as fused coins, regardless of fused
status.

The checkbox's state is per-wallet and is persisted across restarts.

The determination of whether a coin is a fusion coin uses some
heuristics -- it looks for enough ins and outs as well as whether the
special "cashfusion" lokad id is present in the txn.