lnworker: LNWallet.start_network to call super().start_network
The call to super was removed in https://github.com/spesmilo/electrum/commit/4efcb53d24d5f1040556d5fb4194eb8007943602 ,
so that LNWallet's taskgroup would not run _maintain_connectivity, AFAICT.
The way it was done meant that "main_loop" itself would not run for LNWallet, only for LNGossip - very confusing.
It is only due to a quirk in the behaviour of TaskGroups that the group "started" at all.