qt QRDialog: make dialog usefully resizeable
In commit https://github.com/spesmilo/electrum/commit/9bba65199e80556de0c6efd28258344777ac5a6d,
the QRCodeWidget was put inside a BoxLayout as a workaround to avoid the "copy to clipboard" and
"save as file" functionality grabbing extra whitespace/stretch/padding and putting it into the
exported image.

However, in turn that commit introduced a bug, where making the dialog larger does not make the
QRCodeWidget larger (which worked prior).

This commit tries to fix the regression and also the original bug.