document disabling utxo_freeze_threshold
Add documentation to the docstring of the appropriate function
explaining how to disable the #6960 test for dusting transactions

TODO: it may make sense to document passing this option on the command
line as well, since editing the config file by hand is a recipe for
trouble.

Unfortunately this is not as simple as doing:
```
electrum --unconf_utxo_freeze_threshold=0
```
and I can't find how electrum takes config arguments on the command
line. Maybe someone else knows?