qml: in the password dialogs, disable the password confirmation line if the first entered password is too short.
Without that, a user may enter two passwords that are identical
but too short, and then click on the eye icon in order to discover
that they actuall are identical.. and only at this point guess that
the size might be the problem.

Also, raise the minimum length to 6, because that is what is was
on Kivy.

One of the password dialogs still had two eye icons; that was only
fixed in the wizard. I guess that could be avoided if both dialogs
used the same code.