qml: Give user feedback after bumping the fee.
This is better than nothing, but not ideal. This window should
have a general purpose 'userinfo' field, like InvoiceDialog,
that would also display 'Broadcasting...' while the tx is being
broadcast.

Note that in order to bump the fee again, the user will have to
leave this window and open it again.