locale: strip context from strings extracted by lupdate and propagated to gettext by lconvert
This fixes an issue whereby translations that are unique to QML did not get translated because of
a missing context id, not passed to the underlying gettext translator.

As an alternative, we can pass a translation context string from QTranslator to gettext, but that
needs a bit messy mangling of the context string we get from QTranslator.