qml: fix channelopener.connectStr qr scan pop-under
With gossip enabled, trying to open a channel in qml. I try to scan a QR code
of a connect-string (node id), seemingly the scan dialog pops under the
OpenChannelDialog: pointing the device at a QR code results in a successful scan
(so the camera is active despite the image not being shown).

This changes the scanning code in OpenChannelDialog to be similar to other dialogs.